Fix bug where abandoned connections would never idle.
This could happen when query responses were deallocated prior to being fully exhausted. When this happened, the connection was no longer exhaustible and could no longer transition to an idle state as a result.

Instead, we now kill connections when the response has not been fully consumed. Killed connections are removed from the connection pool on subsequent queries.